Title screen
The title screen is packed with stuff, I used some simple animations to fade the links in, I used bootstrap to center the stuff and I used a radial gradient to make the background. I used sass to make the code easier to read and to simplify things and make it easier to write. I also made the links turn gray when hovering over them, added a transition to make it smoother and I made 3 empty html files that will be used as pages in the future. The color and overall design is aimed to be modern and simplistic. For the links I used both grid and flex box to get them centered but also in rows.